+++ DesktopApplicationTemplate.UI/Views/FileObserverCreateServiceView.xaml.cs
@@ -1,0 +1,16 @@
+using System.Windows.Controls;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.UI.ViewModels;
+
+namespace DesktopApplicationTemplate.UI.Views;
+
+public partial class FileObserverCreateServiceView : Page
+{
+    public FileObserverCreateServiceView(FileObserverCreateServiceViewModel vm, ILoggingService logger)
+    {
+        InitializeComponent();
+        DataContext = vm;
+        vm.Logger = logger;
+    }
+}
+